refactor(model): rename the declarative marker namespace to DeclarativeBinding
The Abblix.Oidc.Server.DeclarativeValidation namespace holds ten markers the
source generators read to shape the transport model — only three concern
validation; the rest declare the wire format and the binding source. Rename
it to Abblix.Oidc.Server.DeclarativeBinding to describe the whole set. Both
generators resolve the marker namespace from an anchor, so this rename is
matched, not emitted: the generated output is byte-identical.

BREAKING: the declarative binding markers moved to the
Abblix.Oidc.Server.DeclarativeBinding namespace.
