+++ Abblix.Oidc.Server/DeclarativeBinding/ElementsRequiredAttribute.cs
@@ -1,31 +1,31 @@
 // Abblix OIDC Server Library
 // Copyright (c) Abblix LLP. All rights reserved.
-//
+// 
 // DISCLAIMER: This software is provided 'as-is', without any express or implied
 // warranty. Use at your own risk. Abblix LLP is not liable for any damages
 // arising from the use of this software.
-//
+// 
 // LICENSE RESTRICTIONS: This code may not be modified, copied, or redistributed
 // in any form outside of the official GitHub repository at:
 // https://github.com/Abblix/OIDC.Server. All development and modifications
 // must occur within the official repository and are managed solely by Abblix LLP.
-//
+// 
 // Unauthorized use, modification, or distribution of this software is strictly
 // prohibited and may be subject to legal action.
-//
+// 
 // For full licensing terms, please visit:
-//
+// 
 // https://oidc.abblix.com/license
-//
+// 
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
-namespace Abblix.Oidc.Server.DeclarativeValidation;
+namespace Abblix.Oidc.Server.DeclarativeBinding;
 
 /// <summary>
-/// Declares that the value is the client X.509 certificate presented at the transport layer —
-/// via mutual TLS (RFC 8705) or forwarded by a trusted reverse proxy. Purely semantic: it names
-/// the transport source and leaves the extraction mechanism to the transport layer.
+/// Applied to an array, collection, or otherwise enumerable property, field, or parameter to declare that
+/// the collection itself must be non-null and contain at least one element, and that no element inside it
+/// may be null. Useful for protocol metadata such as <c>redirect_uris</c> where an empty array is invalid.
 /// </summary>
 [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter)]
-public class ClientCertificateAttribute : Attribute;
+public class ElementsRequiredAttribute : Attribute;
