+++ Abblix.Oidc.Server/Endpoints/DynamicClientManagement/Interfaces/ClientRegistrationSuccessResponse.cs
@@ -21,7 +21,7 @@
 // info@abblix.com
 
 using System.Text.Json.Serialization;
-using Abblix.Oidc.Server.DeclarativeValidation;
+using Abblix.Oidc.Server.DeclarativeBinding;
 using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Endpoints.DynamicClientManagement.Interfaces;
