+++ Abblix.Oidc.Server/Model/ClientRegistrationRequest.cs
@@ -24,7 +24,7 @@ using System.Net.Http.Headers;
 using System.Text.Json.Serialization;
 using Abblix.Jwt;
 using Abblix.Oidc.Server.Common.Constants;
-using Abblix.Oidc.Server.DeclarativeValidation;
+using Abblix.Oidc.Server.DeclarativeBinding;
 using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Model;
