+++ Abblix.Oidc.Server/Model/DeviceAuthorizationRequest.cs
@@ -22,7 +22,7 @@
 
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
-using Abblix.Oidc.Server.DeclarativeValidation;
+using Abblix.Oidc.Server.DeclarativeBinding;
 using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Model;
