+++ Abblix.Oidc.Server/Model/ReadClientSuccessfulResponse.cs
@@ -21,7 +21,7 @@
 // info@abblix.com
 
 using System.Text.Json.Serialization;
-using Abblix.Oidc.Server.DeclarativeValidation;
+using Abblix.Oidc.Server.DeclarativeBinding;
 using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Model;
