+++ Abblix.Oidc.Server/Model/TokenRequest.cs
@@ -23,7 +23,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 using Abblix.Oidc.Server.Common.Constants;
-using Abblix.Oidc.Server.DeclarativeValidation;
+using Abblix.Oidc.Server.DeclarativeBinding;
 using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Model;
